# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateListRequest:
    properties:
      description:
        type: string
      game_ids:
        items:
          type: integer
        type: array
      is_public:
        type: boolean
      thumbnail_img:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      field:
        type: string
    type: object
...
  dto.GameDetailResponse:
    properties:
      average_playtime:
        type: number
      avg_rating:
        type: number
      description:
        type: string
      genres:
        items:
          $ref: '#/definitions/dto.GenreDTO'
        type: array
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      is_free:
        type: boolean
      like_count:
        type: integer
      platforms:
        items:
          $ref: '#/definitions/dto.PlatformDTO'
        type: array
      price:
        type: number
      release_date:
        type: string
      review_count:
        type: integer
      steam_id:
        type: integer
      studio:
        $ref: '#/definitions/dto.StudioDTO'
      title:
        type: string
    type: object
...
  dto.GameTrendingResponse:
    properties:
      avg_rating:
        type: number
      game_id:
        type: integer
      release_date:
        type: string
      studios:
        items:
          type: string
        type: array
      thumbnail:
        type: string
      title:
        type: string
      total_reviews:
        type: integer
      trending_score:
        type: integer
    type: object
  dto.GenreDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  dto.LikeGameResponse:
    properties:
      is_liked:
        type: boolean
    type: object
  dto.ListAuthorInfo:
    properties:
      avatar:
        type: string
      username:
        type: string
    type: object
  dto.ListDetailResponse:
    properties:
      author:
        $ref: '#/definitions/dto.ListAuthorInfo'
      created_at:
        type: string
      description:
        type: string
      game_count:
        type: integer
      games:
        items:
          $ref: '#/definitions/dto.ListEntryDTO'
        type: array
      id:
        type: integer
      like_count:
        type: integer
      thumbnail_img:
        type: string
      title:
        type: string
    type: object
  dto.ListEntryDTO:
    properties:
      game_id:
        type: integer
      note:
        type: string
      poster:
        type: string
      title:
        type: string
    type: object
  dto.ListTrendingResponse:
    properties:
      author:
        $ref: '#/definitions/dto.ListAuthorInfo'
      created_at:
        type: string
      game_count:
        type: integer
      list_id:
        type: integer
      thumbnails:
        items:
          type: string
        type: array
      title:
        type: string
      total_likes:
        type: integer
      weekly_likes_count:
        type: integer
    type: object
...
  dto.PaginatedResponse-array_dto_GameTrendingResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.GameTrendingResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationDTO'
      status:
        type: string
    type: object
  dto.PaginatedResponse-array_dto_ListTrendingResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ListTrendingResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationDTO'
      status:
        type: string
    type: object
...
  dto.PaginationDTO:
    properties:
      current_page:
        type: integer
      limit:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
  dto.PlatformDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  dto.RateGameRequest:
    properties:
      game_id:
        type: integer
      rating:
        maximum: 5
        minimum: 0.5
        type: number
    required:
    - game_id
    - rating
    type: object
  dto.RateGameResponse:
    properties:
      my_rating:
        type: number
      new_game_avg:
        type: number
      total_ratings:
        type: integer
    type: object
...
  dto.StudioDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.SuccessResponse-dto_GameDetailResponse:
    properties:
      data:
        $ref: '#/definitions/dto.GameDetailResponse'
    type: object
  dto.SuccessResponse-dto_LikeGameResponse:
    properties:
      data:
        $ref: '#/definitions/dto.LikeGameResponse'
    type: object
  dto.SuccessResponse-dto_ListDetailResponse:
    properties:
      data:
        $ref: '#/definitions/dto.ListDetailResponse'
    type: object
  dto.SuccessResponse-dto_RateGameResponse:
    properties:
      data:
        $ref: '#/definitions/dto.RateGameResponse'
    type: object
...
  dto.SuccessResponse-string:
    properties:
      data:
        type: string
    type: object
  dto.UpdateListRequest:
    properties:
      description:
        type: string
      game_ids:
        items:
          type: integer
        type: array
      is_public:
        type: boolean
      thumbnail_img:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Hệ thống API cho Gaming Box.
  title: Gaming Box API
  version: "1.0"
paths:
...
  /games/{game_id}/like:
    post:
      description: Thích hoặc Bỏ thích một game (Cần đăng nhập)
      parameters:
      - description: ID của Game
        in: path
        name: game_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_LikeGameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Thích hoặc Bỏ thích Game
      tags:
      - Likes
  /games/{id}:
    get:
      description: Lấy thông tin đầy đủ của một game dựa trên ID
      parameters:
      - description: ID của Game
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_GameDetailResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Lấy chi tiết Game
      tags:
      - Games
  /games/popular:
    get:
      description: Lấy danh sách game có đánh giá cao nhất
      parameters:
      - default: 1
        description: Trang hiện tại
        in: query
        name: page
        type: integer
      - default: 12
        description: Số lượng mỗi trang
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_GameTrendingResponse'
      summary: Danh sách Game phổ biến
      tags:
      - Games
  /games/rate:
    post:
      consumes:
      - application/json
      description: Đánh giá sao cho một game (Cần đăng nhập)
      parameters:
      - description: Thông tin đánh giá
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RateGameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_RateGameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Đánh giá Game
      tags:
      - Games
  /games/search:
    get:
      description: Tìm kiếm game theo tên có phân trang
      parameters:
      - description: Từ khóa tìm kiếm
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: Trang hiện tại
        in: query
        name: page
        type: integer
      - default: 12
        description: Số lượng mỗi trang
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_GameTrendingResponse'
      summary: Tìm kiếm Game
      tags:
      - Games
  /games/trending:
    get:
      description: Lấy danh sách game phổ biến có phân trang
      parameters:
      - description: Trang hiện tại (Mặc định 1)
        in: query
        name: page
        type: integer
      - description: Số lượng mỗi trang (Mặc định 12)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_GameTrendingResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Lấy danh sách game thịnh hành
      tags:
      - Games
  /lists:
    post:
      consumes:
      - application/json
      description: Tạo danh sách game cá nhân (Cần đăng nhập)
      parameters:
      - description: Thông tin list
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateListRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_ListDetailResponse'
      security:
      - BearerAuth: []
      summary: Tạo List mới
      tags:
      - Lists
  /lists/{id}:
    delete:
      description: Xóa một danh sách game (Chỉ chủ sở hữu)
      parameters:
      - description: ID của List
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-string'
      security:
      - BearerAuth: []
      summary: Xóa List
      tags:
      - Lists
    get:
      description: Lấy thông tin và danh sách game trong một list
      parameters:
      - description: ID của List
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_ListDetailResponse'
      summary: Xem chi tiết List
      tags:
      - Lists
    put:
      consumes:
      - application/json
      description: Chỉnh sửa thông tin hoặc danh sách game trong list (Chỉ chủ sở
        hữu)
      parameters:
      - description: ID của List
        in: path
        name: id
        required: true
        type: integer
      - description: Thông tin cập nhật
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_ListDetailResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật List
      tags:
      - Lists
  /lists/{list_id}/like:
    post:
      description: Thích hoặc Bỏ thích một list (Cần đăng nhập)
      parameters:
      - description: ID của List
        in: path
        name: list_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_LikeGameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Thích hoặc Bỏ thích List
      tags:
      - Likes
  /lists/trending:
    get:
      description: Lấy danh sách list phổ biến có phân trang
      parameters:
      - description: Trang hiện tại (Mặc định 1)
        in: query
        name: page
        type: integer
      - description: Số lượng mỗi trang (Mặc định 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_ListTrendingResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Lấy danh sách List thịnh hành
      tags:
      - Lists
...
  /reviews/{review_id}/like:
    post:
      description: Thích hoặc Bỏ thích một review (Cần đăng nhập)
      parameters:
      - description: ID của Review
        in: path
        name: review_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_LikeGameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Thích hoặc Bỏ thích Review
      tags:
      - Likes
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"